Temporary workaround to handle `into_iter_err_on_gc_types` (#3)
Upstream has introduced the function `into_iter_err_on_gc_types` which
iterates over core types, but errors on gc types (because they are not
yet implemented). On upstream this means the function will only ever
return `FuncType`s, but we need continuation types as well. So, this
patch introduces a temporary sum type `FuncOrContType` such that we
can return either. Eventually upstream will have to deal with gc
types, and at that point we can revert from this workaround.