Fix validation of continuation instructions (#36)
Validation was broken in the presence of type canonicalisation. It is
important to *only* use canonicalised types. It is currently a little
awkward to work with raw type indices on the continuation
instructions, as they to be canonicalised before use. This patch
implements some helper functions to convert from `HeapType::concrete`
to canonicalised indices.